automata4j-examples: refactored packages to "simple" package
